feat: 패키지 응답 객체와 레포지토리 부터 처리할 객체 Dto 추가

+++ src/main/java/com/wnis/linkyway/dto/PackageResponse.java
@@ -14,5 +14,5 @@ public class PackageResponse {
     private String nickname;
     private Long tagId;
     private String tagName;
-    private Long numberOfCard;
+    private Integer numberOfCard;
 }
